[MOD] changed organisation of the repository, added an arcade namespace, added smart pointers
