+++ include/ICore.hpp
@@ -10,15 +10,19 @@
 
 #include "IEvent.hpp"
 
-class ICore
+namespace arcade
 {
-public:
-    virtual ~ICore() = default;
-    // TODO: Replace raw pointer with reference / smart pointers
-    virtual void manageEvent(IEvent *event) = 0;
 
-protected:
-private:
-};
+    class ICore
+    {
+    public:
+        virtual ~ICore() = default;
+        // TODO: Replace raw pointer with reference / smart pointers
+        virtual void manageEvent(IEvent *event) = 0;
+
+    protected:
+    private:
+    };
+}
 
 #endif /* !ICORE_HPP_ */
