+++ include/components/Vector3D.hpp
@@ -1,0 +1,24 @@
+/*
+** EPITECH PROJECT, 2022
+** dev
+** File description:
+** Vector3D
+*/
+
+#ifndef VECTOR3D_HPP_
+#define VECTOR3D_HPP_
+
+#include "IComponent.hpp"
+
+namespace arcade
+{
+
+    class Vector3D : public IComponent
+    {
+    public:
+        double _x;
+        double _y;
+        double _z;
+    };
+}
+#endif /* !VECTOR3D_HPP_ */
